backend/pyproject.toml
[project]
name = "bootcamp-fastapi-backend"
version = "0.1.0"
description = "FastAPI backend service for bootcamp day1"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]

